rename files, useRevalidator for post page edits
